Refactor error classes into numba.errors
